+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/domaine/department/DepartmentFactoryImpl.java
@@ -1,12 +1,10 @@
 package cm.enspd.educ_connect.domaine.department;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 
 
 @RequiredArgsConstructor
 public class DepartmentFactoryImpl implements DepartmentFactory {
-
   private final DepartmentRepository departmentRepository;
 
   @Override
