+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/domaine/department/program/ProgramData.java
@@ -1,4 +1,4 @@
-package cm.enspd.educ_connect.domaine.program;
+package cm.enspd.educ_connect.domaine.department.program;
 
 import lombok.Builder;
