+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/domaine/user/UserData.java
@@ -1,6 +1,8 @@
 package cm.enspd.educ_connect.domaine.user;
 
 import java.time.LocalDate;
+import java.util.UUID;
+
 import lombok.Builder;
 
 @Builder
@@ -13,4 +15,4 @@ public record UserData(
     Contact contact,
     Training training,
     LocalDate birthdate,
-    int level) {}
+    UUID level) {}
