+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/domaine/user/UserFactoryImpl.java
@@ -1,5 +1,6 @@
 package cm.enspd.educ_connect.domaine.user;
 
+import cm.enspd.educ_connect.domaine.level.LevelId;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -12,7 +13,7 @@ public class UserFactoryImpl implements UserFactory {
         .saveStudent(
             User.builder()
                 .contact(userData.contact())
-                .level(userData.level())
+                .level(new LevelId(userData.level()))
                 .firstName(userData.firstName())
                 .lastName(userData.lastName())
                 .training(userData.training())
