+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/repository/DepartmentSpringRepository.java
@@ -1,7 +1,7 @@
 package cm.enspd.educ_connect.repository;
 
-import cm.enspd.educ_connect.domaine.Department;
-import cm.enspd.educ_connect.domaine.DepartmentId;
+import cm.enspd.educ_connect.domaine.department.Department;
+import cm.enspd.educ_connect.domaine.department.DepartmentId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
