+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/repository/LevelSpringRepository.java
@@ -1,7 +1,7 @@
 package cm.enspd.educ_connect.repository;
 
-import cm.enspd.educ_connect.domaine.Level;
-import cm.enspd.educ_connect.domaine.LevelId;
+import cm.enspd.educ_connect.domaine.level.Level;
+import cm.enspd.educ_connect.domaine.level.LevelId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
