+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/repository/ProgramSpringRepository.java
@@ -1,7 +1,7 @@
 package cm.enspd.educ_connect.repository;
 
-import cm.enspd.educ_connect.domaine.Program;
-import cm.enspd.educ_connect.domaine.ProgramId;
+import cm.enspd.educ_connect.domaine.department.program.Program;
+import cm.enspd.educ_connect.domaine.department.program.ProgramId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
