+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/repository/impl/LevelRepositoryImpl.java
@@ -11,7 +11,7 @@ public class LevelRepositoryImpl implements LevelRepository {
   private final LevelSpringRepository levelSpringRepository;
 
   @Override
-  public Level saveLevel(Level level) {
+  public Level save(Level level) {
     return levelSpringRepository.save(level);
   }
 }
