+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/repository/impl/ProgramRepositoryImpl.java
@@ -1,8 +1,8 @@
 
 package cm.enspd.educ_connect.repository.impl;
 
-import cm.enspd.educ_connect.domaine.program.Program;
-import cm.enspd.educ_connect.domaine.program.ProgramRepository;
+import cm.enspd.educ_connect.domaine.department.program.Program;
+import cm.enspd.educ_connect.domaine.department.program.ProgramRepository;
 import cm.enspd.educ_connect.repository.ProgramSpringRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -11,7 +11,7 @@ public class ProgramRepositoryImpl implements ProgramRepository {
   private final ProgramSpringRepository programSpringRepository;
 
   @Override
-  public Program saveProgram(Program program) {
+  public Program save(Program program) {
     return programSpringRepository.save(program);
   }
 }
