+++ cm.enspd.educ-connect/src/main/java/cm/enspd/educ_connect/service/mapper/LevelMapper.java
@@ -1,6 +1,6 @@
 package cm.enspd.educ_connect.service.mapper;
 
-import cm.enspd.educ_connect.domaine.LevelData;
+import cm.enspd.educ_connect.domaine.level.LevelData;
 import cm.enspd.educ_connect.dto.LevelDTO;
 import org.mapstruct.BeanMapping;
 import org.mapstruct.InjectionStrategy;
@@ -11,6 +11,5 @@ import org.mapstruct.Mapping;
 public interface LevelMapper {
     @BeanMapping(ignoreByDefault = true)
     @Mapping(target = "name")
-    @Mapping(target = "trainingType")
     LevelData mapToLevelData(LevelDTO levelDTO);
 }
